feat(event-runtime): shared RunDetailBlocks — one run-detail renderer, tiered fields (WM-129)
Extract the run detail blocks duplicated between the Runs side panel and
the #/run full page into components/RunDetailBlocks.tsx, so the two views
cannot drift (the full page had lost the Deadlines clocks, placement,
attempt timestamps, and gained divergent styling).

Tier the Run section for glanceability: spec.input promoted to
first-class rows (the "what is this run working on?" answer used to be
buried in the collapsed RunSpec JSON); run id, idempotencyKey, specHash,
inputHash, workspace path, prompt/policy versions demoted to a collapsed
internals disclosure. Nothing removed, only demoted; the immutable
RunSpec JSON stays as ground truth.
